final changes

+++ flask-sqlite3/api.py
@@ -170,19 +170,28 @@ def removeFollower():
     checkUserQuery = """SELECT id, username FROM users WHERE username=?"""
     userExistData = (userName,)
     user_result = query_db_check(checkUserQuery, userExistData)
+    print(user_result)
     userExistData = (userNameToUnFollow,)
     follow_user_result = query_db_check(checkUserQuery, userExistData)
+    checkfollowingQuery = """Select userId,following from followers where userid in (select userId from users where userName = ?) and following in (select id from users where userName = ?)"""
+    checkFollowingData = (userName,userNameToUnFollow)
+    checkExistence = query_db_check(checkfollowingQuery,checkFollowingData)
+
     if user_result and follow_user_result:
-        sql_select = """Select id from users where userName = ?"""
-        data = (userNameToUnFollow,)
-        idOfFollowing = query_db_check(sql_select, data).get("id")
-        data = (userName,)
-        idOfUser = query_db_check(sql_select, data).get("id")
-        sql_delete = """DELETE from followers where userId = ? and following = ?"""
-        values = (idOfUser, idOfFollowing)
-        query_db(sql_delete, values)
-        message = str(userName + ' has stopped following ' + userNameToUnFollow)
-        return {'message': message, 'statueCode': 201}
+        if checkExistence :
+            sql_select = """Select id from users where userName = ?"""
+            data = (userNameToUnFollow,)
+            idOfFollowing = query_db_check(sql_select, data).get("id")
+            data = (userName,)
+            idOfUser = query_db_check(sql_select, data).get("id")
+            sql_delete = """DELETE from followers where userId = ? and following = ?"""
+            values = (idOfUser, idOfFollowing)
+            query_db(sql_delete, values)
+            message = str(userName + ' has stopped following ' + userNameToUnFollow)
+            return {'message': message, 'statueCode': 201}
+        else:
+            message = str(userName + ' does not follows '+userNameToUnFollow)
+            make_error(400, message)
     else:
         make_error(400, 'user Or UserToFollow Does Not Exists')
